requirements.txt
click==8.0.4
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
SQLAlchemy==1.4.32
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
